laravel 8 migration

+++ app/Http/Controllers/admin/ProductController.php
@@ -1,53 +1,41 @@
 <?php
-
 namespace App\Http\Controllers\admin;
 
-use App\Http\Controllers\Controller;
 use App\Product;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\CreateProduct;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Database\Eloquent\Builder;
 
 class ProductController extends Controller
 {
-    public function __construct(){
-        $this->middleware('auth:admin');
-    }
-    
-    public function product(){
-        $name = isset($_REQUEST['name']) ? trim($_REQUEST['name']) : "";
-        $status = isset($_REQUEST['status']) ? trim($_REQUEST['status']) : "";
+    public function product(Request $request){
+        abort_unless($this->checkPermission('View Product'), 403);
         $query = Product::whereNull('deleted_at')->orderBy('id', 'desc');
-        if(!empty($name)){
-            $query->where('product_name','LIKE','%'.$name.'%');
+        if(!empty($request->input('name'))){
+            $query->where('product_name','LIKE','%'. $request->input('name').'%');
         }
-        if(!empty($status)){
-            $query->where('status',$status);
+        if(!empty($request->input('status'))){
+            $query->where('status', $request->input('status'));
         }
-        $product = $query->paginate(10);
-        $data['pageData'] = $product;
+        $data['pageData'] = $query->paginate(10);
         $data['pageTitle'] = "Product";
         return view('admin.product.index')->with('data',$data);
     }
 
     public function create_product(){
-        $data['action'] = "Add";        
-        $data['pageTitle'] = "Add Catgory";
+        abort_unless($this->checkPermission('Create Product'), 403);
+        $data['action'] = "Add";
+        $data['pageTitle'] = "Add Product";
         $data['pageData']['category'] = Builder::getCategoryData();
         return view('admin.product.product_action')->with('data',$data);
     }
-    
-    public function store_product(Request $request){
-        $validated = $request->validate([
-            'category_id' => 'required|integer',
-            'subcategory_id' => 'required|integer',
-            'product_name' => 'required|max:255',
-            'product_description' => 'required|max:255',
-            'price' => 'required|integer',
-            'status' => 'required',
-            'product_image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-        ]);
-        //Uploading Image
+
+    public function store_product(CreateProduct $request){
+        abort_unless($this->checkPermission('Create Product'), 403);
+
+        $newFileName = "";
         if ($request->hasFile('product_image')) {
             $extension = $request->file('product_image')->extension();
             $newFileName = "PRODUCT_".time().".".$extension;
@@ -56,8 +44,6 @@ class ProductController extends Controller
                 Storage::makeDirectory($uploadPath);
             }
             $request->file('product_image')->storeAs($uploadPath,$newFileName);
-        }else{
-            $newFileName = "";
         }
 
         $Product = new Product();
@@ -66,54 +52,31 @@ class ProductController extends Controller
         $Product::Create($input);
         return redirect()->route('admin.product')->with('success','Data Updated Successfuly');
     }
-    
+
     public function edit_product($id){
+        abort_unless($this->checkPermission('Edit Product'), 403);
         $data['action'] = "Edit";
         $data['pageData'] = Product::find($id);
         $data['pageTitle'] = "Edit Product";
         $data['pageData']['category'] = Builder::getCategoryData();
         return view('admin.product.product_action')->with('data',$data);
     }
-    
-    public function update_product($id, Request $request){
-        $validated = $request->validate([
-            'category_id' => 'required|integer',
-            'subcategory_id' => 'required|integer',
-            'product_name' => 'required|max:255',
-            'product_description' => 'required|max:255',
-            'price' => 'required|integer',
-            'status' => 'required',
-            'product_image' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-        ]);
-        $Product = Product::find($id);
-        //Uploading Image
-        if ($request->hasFile('product_image')) {
-            //--------------Method1--------------
-            //$path = $request->file('product_image')->store('/public/productimg');
-            //--------------Method2--------------
-            $extension = $request->file('product_image')->extension();
-            $newFileName = "PRODUCT_".time().".".$extension;
-            $uploadPath = '/public/product/';
-            if(!Storage::exists($uploadPath)){
-                Storage::makeDirectory($uploadPath);
-            }
-            if(Storage::exists($uploadPath.$Product->product_image)) {
-                Storage::delete($uploadPath.$Product->product_image);
-            }
-            $request->file('product_image')->storeAs($uploadPath,$newFileName);
-        }else{
-            $newFileName = $Product->product_image;
-        }
-        $input = $request->all();
-        $input['product_image'] = $newFileName;        
-        $Product->fill($input)->save();
-        return redirect()->route('admin.product')->with('success','Data Updated Successfuly');
+
+    public function update_product($id, CreateProduct $request){
+        abort_unless($this->checkPermission('Edit Product'), 403);
+        $product = Product::find($id);
+        $update = $product->fill($request->all())->save();
+        if($update):
+            return redirect()->route('admin.product')->with('success','Data Updated Successfuly');
+        else:
+            return redirect()->route('admin.product')->with('error', 'Data Failed to update');
+        endif;
     }
-    
+
     public function destroy_product( Request $request){
+        abort_unless($this->checkPermission('Delete Product'), 403);
         $product = Product::find($request->dataId);
         $product->delete();
         echo 1;
     }
-
 }
